refactor: 멀티파트 요청 함수화
- Data의 extension으로 함수화해서 다른 사람도 쉽게 사용할 수 있도록 변경
